Move :compose:icons into sdk
