+++ sdk/compose/icons/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/compose/icons/ValkyrieIcons.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.compose.icons
+package io.github.composegears.valkyrie.sdk.compose.icons
 
 object ValkyrieIcons {
     object Outlined
