+++ tools/compose-app/build.gradle.kts
@@ -38,7 +38,7 @@ kotlin {
 
         commonMain.dependencies {
             implementation(projects.sdk.compose.foundation)
-            implementation(projects.compose.icons)
+            implementation(projects.sdk.compose.icons)
             implementation(projects.shared)
 
             implementation(compose.material3)
