+++ tools/idea-plugin/build.gradle.kts
@@ -54,9 +54,9 @@ dependencies {
     implementation(projects.components.generator.iconpack)
     implementation(projects.components.generator.jvm.imagevector)
     implementation(projects.components.parser.unified)
-    implementation(projects.compose.icons)
-    implementation(projects.sdk.compose.highlightsCore)
     implementation(projects.sdk.compose.foundation)
+    implementation(projects.sdk.compose.highlightsCore)
+    implementation(projects.sdk.compose.icons)
     implementation(projects.sdk.core.extensions)
     implementation(projects.sdk.intellij.psi.iconpack)
     implementation(projects.sdk.intellij.psi.imagevector)
